requirements.txt
openai_whisper==20250625
torch==2.8.0
